# agentpit/openapi.yaml
openapi: 3.1.0
info:
  title: Magic Mirror Agent API
  version: 1.0.0
  description: AgentPit 对接用智能体接口，网关转发到本地 Ollama。
servers:
  - url: https://law.babelbeast.com
paths:
...
  /api/v1/tokens/report:
    get:
      summary: 查询最近 Token 上报记录
      description: 返回按创建时间倒序的最近上报记录
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: 查询成功
    post:
      summary: Token 消耗上报
      description: 使用 ApiKey Bearer Token 上报调用 token 消耗
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - agentId
                - tokensUsed
                - startedAt
                - endedAt
              properties:
                agentId:
                  type: string
                tokensUsed:
                  type: integer
                inputTokens:
                  type: integer
                outputTokens:
                  type: integer
                startedAt:
                  type: string
                  format: date-time
                endedAt:
                  type: string
                  format: date-time
                modelName:
                  type: string
                requestId:
                  type: string
                metadata:
                  type: object
      responses:
        "200":
          description: 上报成功
        "400":
          description: 参数错误
        "401":
          description: 认证失败